build-your-first-image: Fix "Sign the model" hyperlink
I noticed in Build your first image[1] the "Sign the model" hyperlink[2] does not actually go to the appropriate section; it goes to the top.

That is because the page *and* section are both named "Sign the model":

```
1 # Sign the model
[...]
69 ## Sign the model
```

I'm not sure how this can be fixed properly. For now, I'm making it hyperlink to "id1"[3] as I see on my browser.

[1]: https://documentation.ubuntu.com/core/tutorials/build-your-first-image/index.html
[2]: https://documentation.ubuntu.com/core/tutorials/build-your-first-image/sign-the-model/#sign-the-model
[3]: https://documentation.ubuntu.com/core/tutorials/build-your-first-image/sign-the-model/#id1